+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - NEUCORE_MUMBLE_PLUGIN_DB_USERNAME and NEUCORE_MUMBLE_PLUGIN_DB_PASSWORD are now optional: user and password
   can be added to NEUCORE_MUMBLE_PLUGIN_DB_DSN.
+- Removed unused config values from mumble-authenticator.ini.
 
 ## 0.2.0, 2023-01-22